2017.06.03 These changes apply to the R workflow, and have been done to speed up the C targets. To do this the mapping routine was eliminated for this target. In order to do this the calculation of secondary parameters, outputs, etc. was moved from an R script directly into the compiled C library.
The mapped parameters and secondary parameters were subsequently moved from
the output in som$meta to columns in som$simout. This might break some scripts
in place from users, but it's such a huge speed improvement that it's worth
it. Testing has been done to the following ubiquity R functions:

 - simulate_subjects
 - run_simulation_titrate
 - estimate_parameters
 - som_to_df

And the Shiny App (model_gui.r). It all seems to be working but there are
probably some things that have been broken that will pop up later.

In addition to the stuff found in som$meta that was move to som$simout. The
simout will also contain all of the secondary parameters (dynamic and static),
covariates (initial condition and timecourse), infusino rate timecourse, etc.
This can now be used by users for debugging models.