2017.07.10
Added the ability to read subject level parameters and covariates from a file
when performing stochastic simulations in the R workflow.

Added error checking for analysis names and display errors to users if there
are problems.

Added more verbose output when starting an estimation

Fixed documentation in matlab for some functions that are out of date.

Small bug fixes in parameter estimation in Matlab.

Fixed BIC/AIC calculations in R workflow for maximum-likelihood estimation.