R Workflow
Added saving of sessionInfo() and update strings for the system file wile
performing estimations.

Trying to fix errors like the follwoing that seem to have been introduced with
R 4.0:

Error in if (!file.exists(cfg$options$logging$file)) { (from ubiquity.R#3932) :
  argument is of length zero

Replaced all of the references in the rcomponents file using the dollar
operator with the [[]] operators.